plugin installation is now part of the CLI
