+++ bin/kuzzle-enable.js
