+++ bin/kuzzle-start.js
@@ -3,6 +3,15 @@ var
   rc = require('rc'),
   kuzzle = require('../lib');
 
+if (process.env.NEW_RELIC_APP_NAME) {
+  require('newrelic');
+}
+
+if (process.env.FEATURE_COVERAGE == 1) {
+  var coverage = require('istanbul-middleware');
+  console.log('Hook loader for coverage - ensure this is not production!');
+  coverage.hookLoader(__dirname+'/lib');
+}
 
 module.exports = function () {
   console.log('Starting Kuzzle');
