+++ lib/api/core/pluginsManager.js
@@ -1,16 +1,13 @@
 var
-  InternalError = require('./errors/internalError'),
   GatewayTimeoutError = require('./errors/gatewayTimeoutError'),
   PluginContext = require('./pluginsContext'),
   fs = require('fs'),
-  q = require('q'),
   async = require('async'),
   path = require('path'),
-  _ = require('lodash'),
   childProcess = require('child_process'),
-  pathConfig = path.join(__dirname, '..', '..', '..', 'config'),
-  lockfile = require('proper-lockfile');
-
+  q = require('q'),
+  _ = require('lodash'),
+  pathConfig = path.join(__dirname, '..', '..', '..', 'config');
 
 /*
  We use the console to display information, as there may be no logger plugin available while installing/launching
@@ -34,68 +31,32 @@ module.exports = function PluginsManager (kuzzle) {
    *
    * @param {Boolean} isServer, true if this is a server instance, false for worker instances
    * @param {Boolean} isDummy, true if we are trying to test pluginsManager
-   * @returns {Promise}
    */
   this.init = function (isServer, isDummy) {
     var
       pathDefaultPlugins = path.join(pathConfig, 'defaultPlugins.json'),
       pathCustomPlugins = path.join(pathConfig, 'customPlugins.json'),
-      defaultPlugins,
-      updateInProgress = false,
-      customPlugins = {},
-      deferred;
+      defaultPlugins = require(pathDefaultPlugins),
+      customPlugins = {};
+
+    if (!childProcess.hasOwnProperty('execSync')) {
+      console.error('███ kuzzle-install: Make sure you\'re using Node version >= 0.12');
+      process.exit(1);
+    }
 
     this.isDummy = isDummy;
     this.isServer = isServer;
 
     if (this.isDummy) {
-      return q({});
+      return false;
     }
 
-    if (!childProcess.hasOwnProperty('execSync')) {
-      console.error('Make sure you\'re using Node version >= 0.12');
-      process.exit(1);
+    if (fs.existsSync(pathCustomPlugins)) {
+      customPlugins = require(pathCustomPlugins);
     }
 
-    deferred = q.defer();
-
-    /*
-     Prevents multiple plugin installations at the same time.
-     Delay installation for an instance only if it's installed on the same computer than another one.
-
-     Skip plugin installation if we're locked, as it means that another instance is already updating plugins
-     */
-    updateInProgress = fs.existsSync('./node_modules.lock');
-
-    lockfile.lock('./node_modules', {retries: 1000, minTimeout: 200, maxTimeout: 1000}, (err, release) => {
-      if (err) {
-        return deferred.reject(err);
-      }
-
-      defaultPlugins = require(pathDefaultPlugins);
-
-      if (fs.existsSync(pathCustomPlugins)) {
-        customPlugins = require(pathCustomPlugins);
-      }
-
-      if (!updateInProgress) {
-        if (installPlugins.call(this, defaultPlugins)) {
-          fs.writeFileSync(pathDefaultPlugins, JSON.stringify(defaultPlugins, null, 2));
-        }
-
-        if (installPlugins.call(this, customPlugins, defaultPlugins)) {
-          fs.writeFileSync(pathCustomPlugins, JSON.stringify(customPlugins, null, 2));
-        }
-      }
-
-      release();
-
-      this.plugins = constructList(defaultPlugins, customPlugins, this.isServer);
-      loadPlugins(this.plugins);
-      deferred.resolve();
-    });
-
-    return deferred.promise;
+    this.plugins = constructList(defaultPlugins, customPlugins, this.isServer);
+    loadPlugins(this.plugins);
   };
 
   /**
@@ -286,47 +247,6 @@ function triggerPipes(event, data) {
   return deferred.promise;
 }
 
-/**
- * Install given plugins
- * @param plugins
- * @param basePlugins
- * @returns {boolean}
- */
-function installPlugins(plugins, basePlugins) {
-  var
-    newInstalled = false,
-    pluginInstallId;
-
-  _.forEach(plugins, (plugin, name) => {
-    if (plugin.url) {
-      pluginInstallId = plugin.url;
-    }
-    else if (plugin.version) {
-      pluginInstallId = name + '@' + plugin.version;
-    }
-    else {
-      console.error('Plugin', name, 'has no version. The version is mandatory if there is no URL.');
-      return true;
-    }
-
-    if (!needInstall(name, pluginInstallId)) {
-      return true;
-    }
-
-    newInstalled = true;
-    npmInstall(pluginInstallId);
-    initConfig.call(this, plugin, name);
-
-    // By default, when a new plugin is installed, the plugin is disabled
-    // If in customPlugins the `activated` flag is undefined, we get the `activated` flag from the default one
-    if (plugin.activated === undefined) {
-      plugin.activated = (basePlugins !== undefined && basePlugins[name] && basePlugins[name].activated === true);
-    }
-  });
-
-  return newInstalled;
-}
-
 function constructList(defaultPlugins, customPlugins, isServer) {
   var
     allPlugins = _.extend(defaultPlugins, customPlugins),
@@ -377,91 +297,6 @@ function constructList(defaultPlugins, customPlugins, isServer) {
   });
 
   return pluginsList;
-}
-
-/**
- * Execute shell command
- * @param command
- */
-function sh(command) {
-  return childProcess.execSync(command).toString();
-}
-
-/**
- * Install a plugin with NPM
- * @param plugin
- */
-function npmInstall(plugin) {
-  sh('npm install ' + plugin);
-}
-
-/**
- * Initialize the config plugin
- * @param plugin
- * @param name
- * @returns {boolean}
- */
-function initConfig(plugin, name) {
-  var
-    pluginPackage;
-
-  try {
-    pluginPackage = require(path.join(getPathPlugin(name), 'package.json'));
-  }
-  catch (e) {
-    console.error(new InternalError('There is a problem with plugin ' + name + '. Check the plugin name'));
-  }
-
-  // If there is no information about plugin in the package.json
-  if (!pluginPackage.pluginInfo) {
-    return false;
-  }
-
-  plugin = _.extend(plugin, pluginPackage.pluginInfo);
-}
-
-/**
- * Function for detect if the configured plugin must be installed
- * If the plugin is configured with an url from GIT, the plugin is installed each time
- * If the plugin come from NPM, the plugin is installed only if the version is different from the already installed
- *
- * @param name
- * @param from previously installation information with version or git url with branch
- * @returns {boolean} true if the plugin must be installed, false if not
- */
-function needInstall(name, from) {
-  var
-    packageDefinition,
-    packagePath,
-    pluginPath = getPathPlugin(name);
-
-  // If we want to install a plugin with git, maybe there is no version and we want to 'pull' the plugin
-  if (from.indexOf('git') !== -1) {
-    return true;
-  }
-
-  if (!fs.existsSync(pluginPath)) {
-    return true;
-  }
-
-  packagePath = path.join(pluginPath, 'package.json');
-  if (!fs.existsSync(packagePath)) {
-    return true;
-  }
-
-  packageDefinition = require(path.join(pluginPath, 'package.json'));
-
-  // If version in package.json is different from the version the plugins.json, we want to install the updated plugin
-  return (packageDefinition._from !== from);
-}
-
-/**
- * Return the real plugin path
- * @param name
- * @returns {String}
- */
-function getPathPlugin (name) {
-  return path.join(__dirname, '..', '..', '..', 'node_modules', name);
 }
 
 function loadPlugins(plugins) {
