+++ lib/api/start.js
@@ -55,70 +55,67 @@ module.exports = function start (params, feature) {
   this.config.version = packageInfo.version;
 
   this.pluginsManager = new PluginsManager(this);
-  this.pluginsManager.init(this.isServer, feature.dummy)
-    .then(() => {
-      this.pluginsManager.run();
+  this.pluginsManager.init(this.isServer, feature.dummy);
+  this.pluginsManager.run();
 
-      if (!this.isWorker) {
-        if (!feature.dummy) {
-          this.services.init({server: true});
+  if (!this.isWorker) {
+    if (!feature.dummy) {
+      this.services.init({server: true});
 
-          this.services.list.readEngine.listIndexes(new RequestObject({}))
-            .then(result => {
-              result.data.indexes.forEach(index => {
-                this.indexes[index] = [];
+      this.services.list.readEngine.listIndexes(new RequestObject({}))
+        .then(result => {
+          result.data.indexes.forEach(index => {
+            this.indexes[index] = [];
 
-                this.services.list.readEngine.listCollections(new RequestObject({index: index}))
-                  .then(resultCollections => {
-                    this.indexes[index] = resultCollections.data.collections;
-                  });
+            this.services.list.readEngine.listCollections(new RequestObject({index: index}))
+              .then(resultCollections => {
+                this.indexes[index] = resultCollections.data.collections;
               });
-            });
-        }
-        else {
-          this.services.init({server: false, blacklist: ['mqBroker', 'perf', 'writeEngine', 'readEngine', 'notificationCache', 'monitoring', 'remoteActions', 'statsCache']});
-        }
+          });
+        });
+    }
+    else {
+      this.services.init({server: false, blacklist: ['mqBroker', 'perf', 'writeEngine', 'readEngine', 'notificationCache', 'monitoring', 'remoteActions', 'statsCache']});
+    }
 
-        // The funnel controller dispatch messages between the router controller and other controllers
-        this.funnel = new FunnelController(this);
-        this.funnel.init();
+    // The funnel controller dispatch messages between the router controller and other controllers
+    this.funnel = new FunnelController(this);
+    this.funnel.init();
 
-        // The router controller listens to client requests and pass them to the funnel controller
-        this.router = new RouterController(this);
+    // The router controller listens to client requests and pass them to the funnel controller
+    this.router = new RouterController(this);
 
-        // Room subscriptions core components
-        this.hotelClerk = new HotelClerk(this);
-        this.dsl = new Dsl(this);
+    // Room subscriptions core components
+    this.hotelClerk = new HotelClerk(this);
+    this.dsl = new Dsl(this);
 
-        // Notifications core component
-        this.notifier = new Notifier(this);
-        this.notifier.init();
+    // Notifications core component
+    this.notifier = new Notifier(this);
+    this.notifier.init();
 
-        // Statistics core component
-        this.statistics = new Statistics(this);
+    // Statistics core component
+    this.statistics = new Statistics(this);
 
-        // Worker response listener
-        this.workerListener = new WorkerListener(this, this.config.queues.workerWriteResponseQueue);
+    // Worker response listener
+    this.workerListener = new WorkerListener(this, this.config.queues.workerWriteResponseQueue);
 
-        if (!feature.dummy) {
-          // Initialize hooks
-          this.hooks.init();
+    if (!feature.dummy) {
+      // Initialize hooks
+      this.hooks.init();
 
-          // Starts the servers in charge of listening to client queries (HTTP, MQ or WebSocket)
-          servers.initAll(this, params);
-        }
-      }
+      // Starts the servers in charge of listening to client queries (HTTP, MQ or WebSocket)
+      servers.initAll(this, params);
+    }
+  }
 
-      // Start a single set of workers
-      if (this.isWorker === undefined || this.isWorker === true) {
-        this.workers.init();
-      }
+  // Start a single set of workers
+  if (this.isWorker === undefined || this.isWorker === true) {
+    this.workers.init();
+  }
 
-      // the repositories need to be instanciated after the services are initialized
-      this.repositories = require('./core/models/repositories')(this);
-      kuzzleStarted.resolve({});
-    })
-    .catch(err => kuzzleStarted.reject(err));
+  // the repositories need to be instanciated after the services are initialized
+  this.repositories = require('./core/models/repositories')(this);
+  kuzzleStarted.resolve({});
 
   return kuzzleStarted.promise;
 };
